Task 3

+++ csharp-linear_algebra/3-magnitude/3-magnitude.cs
@@ -12,12 +12,12 @@ class VectorMath
 
         if (c == 2)
         {
-            res = Math.Round(Math.Sqrt(Math.Pow(vector[0], 2) + Math.Pow(vector[1], 2)), 2);
+            res = (int)Math.Round(Math.Sqrt(Math.Pow(vector[0], 2) + Math.Pow(vector[1], 2)), 2);
             return(res);
         }
         else if (c == 3)
         {
-            res = Math.Round(Math.Sqrt(Math.Pow(vector[0], 2) + Math.Pow(vector[1], 2) + Math.Pow(vector[2], 2)), 2);
+            res = (int)Math.Round(Math.Sqrt(Math.Pow(vector[0], 2) + Math.Pow(vector[1], 2) + Math.Pow(vector[2], 2)), 2);
             return(res);
         }
         else
